Azurefunctiondocumentation/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
azure-ai-formrecognizer==3.3.3
azure-communication-email
azure-ai-documentintelligence==1.0.2
azure-common==1.1.28
azure-core==1.35.0
azure-cosmos==4.9.0
azure-functions==1.23.0
azure-identity==1.19.0
azure-storage-blob==12.24.1
azurefunctions-extensions-base==1.0.0
azurefunctions-extensions-bindings-blob==1.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
dataclasses-json==0.6.7
distro==1.9.0
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
frozenlist==1.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-modelarmor==0.2.8
google_search_results==2.4.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-google-community==2.0.7
langchain-google-genai==2.1.10
langchain-openai==0.3.32
langchain-text-splitters==0.3.10
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.7
langsmith==0.4.20
MarkupSafe==3.0.2
marshmallow==3.26.1
msal==1.33.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.2
oauthlib==3.3.1
openai==1.102.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
reportlab==4.4.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0